backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
langgraph==0.0.40
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.10
langchain-ollama==0.1.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
openai==1.3.7
